+++ .env
@@ -2,3 +2,6 @@ CLIENT_ID=368334907870-fbjd0e76huoaje2jfv9156jq6q7842cv.apps.googleusercontent.c
 CLIENT_SECRET=0GxO13B6ZcUeKWy2jh-e_x17
 OUTLOOK_CLIENT_ID=bae8a641-8157-4d9c-a6da-bb2673a8d544
 OUTLOOK_CLIENT_SECRET=JQL84Ep5QigwTN5MU-~T9GT8.o4Ai6qZ~i
+
+MONGO_URI = mongodb+srv://admin-rabvik:Test123@cluster0.thhlu.mongodb.net/userDB
+MONGO_URI_LOCAL = mongodb://localhost/userDB
