+++ README.md
@@ -5,7 +5,7 @@
     ```
 - start the development server
     ```
-    npm start
+    npm run devStart
     ```
     Runs the app in the development mode.\
     Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
