+++ public/assets/js/generic.js
@@ -2,123 +2,7 @@ const form = document.querySelector('.form');
 form.addEventListener('submit',(e)=>{
     // e.preventDefault();
 })
-const sectorOptions = [
-    '3d printing','Accounting','AdTech','Advisory','Agri-Tech','Agricultural Chemicals',
-    'Agriculture','Animal Husbandry','Apparel','Apparel & Accessories','Application Development',
-    'Art','Auto & Truck Manufacturers','Auto Vehicles', 'Parts & Service Retailers',
-    'Auto, Truck & Motorcycle Parts','Aviation & Others','Baby Care','Big Data','Billing and Invoicing',
-    'Biotechnology','Bitcoin and Blockchain','BPO','Branding','Business Finance','Business Intelligence',
-    'Business Support Services','Business Support Supplies','Clean Tech','Clean Technology','Cloud',
-    'Coaching','Collaboration','Commercial Printing Services','Commodity Chemicals','Comparison Shopping',
-    'Computer & Electronics','Retailers','Construction & Engineering','Construction Materials',
-    'Construction Supplies & Fixtures','Consulting','Consumer Products','Consumer Services',
-    'Corporate Social','Responsibility','Coworking Spaces','Crowdfunding','Customer Support','CXM',
-    'Cyber Security','Dairy Farming','Data Science','Defence Equipment','Digital Marketing',
-    'Digital Marketing (SEO Automation)','Digital Media','Digital Media Blogging',
-    'Digital Media NewsDigital Media','Publishing','Digital Media Video','Discovery',
-    'Diversified Chemicals','Drones','E-Commerce','E-learning','Education','Education Technology',
-    'Electric Vehicles','Electronics','Electronics / Instrumentation','Embedded','Employment Services',
-    'Enterprise Mobility','Entertainment','Environmental Services & Equipment','ERP','Event Management',
-    'Experiential Travel','Facility Management','Fan Merchandise','Fantasy Sports','Fashion',
-    'Fashion Technology','Fintech','Fisheries','Food and Beverage','Food Processing',
-    'Food Technology/Food Delivery','Foreign Exchange','Freight & Logistics Services','Handicraft',
-    'Health & Wellness','Healthcare IT','Healthcare Services','Healthcare Technology','Holiday Rentals',
-    'Home Care','Home Furnishings Retailers','Home Improvement Products & Services','Retailers',
-    'Home Security solutions','Homebuilding','Horticulture','Hospitality','Hotel','Housing',
-    'Industrial Design','Industrial/Energy','Insurance','Integrated communication services',
-    'Internet / Web Services','Internships','IT Consulting','IT Management','IT Services','JewelleryKPO',
-    'Laundry','Leather Footwear','Leather Textiles Goods','Legal','Lifestyle','Location-Based','Loyalty',
-    'Machine Learning','Manufacture of Electrical','Equipment','Manufacture of Machinery and Equipment',
-    'Manufacturing','Manufacturing & Warehouse','Market Research','Marketing / Advertising',
-    'Media and Entertainment','Medical Devices and Equipment','Medical Devices Biomedical','Microbrewery',
-    'Microfinance','Mobile wallets Payments','Movies','Natural Language Processing',
-    'Network Technology Solutions','New-age Construction','Technologies','NGO','NLP','Non-Leather Footwear',
-    'Non-Leather Textiles Goods','Oil & Gas Drilling','Oil & Gas Exploration and Production',
-    'Oil & Gas Transportation Services','Oil Related Services and','Equipment','Online Classified',
-    'OOH Media','Organic Agriculture','Others','P2P Lending','Passenger Transportation Services',
-    'Payment Platforms','Personal Care','Personal Finance','Personal Security','Pharmaceutical',
-    'Photography','Physical Toys and Games','Point of Sales','Product Development',
-    'Professional Information Services','Project Management','Public Citizen Security Solutions',
-    'Recruitment Jobs','Renewable Energy Solutions','Renewable Nuclear Energy','Renewable Solar Energy',
-    'Renewable Wind Energy','Restaurants','Retail Technology','Robotics Application','Robotics Technology',
-    'Sales','SCM','Semiconductor','Skill Development','Skills Assessment','Smart Home','Social Commerce',
-    'Social Media','Space Technology','Specialty Chemicals','Sports Promotion and Networking',
-    'Talent Management','Testing','Textile','Ticketing','Tires & Rubber Products','Trading',
-    'Traffic Management','Training','Transport Infrastructure','Transportation','Travel','Utility Services',
-    'Virtual Games','Waste Management','Wayside Amenities','Wearables','Web Design','Web Development',
-    'Weddings','Wireless','Other'
-];
-const businessActivities = [
-    'e-Shop','e-Mall','e-Auction','Virtual community','Collaborative platform','Marketplace',
-    'Added value service','Infrastructure portal','Infrastructure producer','Infomediary',
-    'Affiliate','Direct sales','Other'
-]
 
-const marketCategory = ['B2C','B2B','B2B2C','Other']
-
-const devStages = ['Ideation','Concept','Validation','Early Traction','Scaling','Growth']
-
-const services = ['IaaS platform','SaaS software','IoT','PaaS platform','Business Process as a Service (BPaaS)',
-    'Hardware','Software On-Premises','Intellectual services','Other']
-
-function addOptions(select,value){
-    let option = document.createElement('option');
-    option.innerHTML = value;
-    option.value = value.toLowerCase();
-    select.appendChild(option);
-}
-
-//------------sector-----------------------
-const sectorSelect = document.querySelector('.sector-select')
-let selectedSector;
-sectorSelect.addEventListener('change',(e)=>{
-    selectedSector = e.target.value;
-    console.log(selectedSector);
-})
-sectorOptions.forEach(sector=>{
-    addOptions(sectorSelect,sector);
-});
-
-//---------busness activity----------------------
-const businessSelect = document.querySelector('.business-select')
-let selectedBusiness;
-businessSelect.addEventListener('change',(e)=>{
-    selectedBusiness = e.target.value;
-    console.log(selectedBusiness);
-})
-businessActivities.forEach(business=>{
-    addOptions(businessSelect,business);
-});
-
-//-----------------market category------------------------
-const marketSelect = document.querySelector('.market-select');
-let selectedMarket;
-marketSelect.addEventListener('change',(e)=>{
-    selectedMarket = e.target.value;
-});
-marketCategory.forEach(market=>{
-    addOptions(marketSelect,market)
-})
-
-//---------------- stage of development------------
-const stageSelect = document.querySelector('.stage-select');
-let selectedStage;
-stageSelect.addEventListener('change',e=>{
-    selectedStage = e.target.value;
-});
-devStages.forEach(stage=>{
-    addOptions(stageSelect,stage)
-});
-
-//-------------------product/service------------
-const serviceSelect = document.querySelector('.service-select');
-let selectedService;
-serviceSelect.addEventListener('change',e=>{
-    selectedService = e.target.value;
-});
-services.forEach(service=>{
-    addOptions(serviceSelect,service)
-})
 
 //-------add founder list-----------------------
 const founderList = document.querySelector('.founder-list');
@@ -144,7 +28,7 @@ addEmployeeBtn.addEventListener('click',()=>{
         return
     }
     const employeeItem = document.createElement('li');
-    employeeItem.innerHTML = employeeInput.value;
+    employeeItem.innerHTML = `${employeeInput.value}<input value=${employeeInput.value} type='hidden' name='employees[]' />`;
     employeeList.appendChild(employeeItem);
     employeeInput.value='';
 });
