+++ views/questionnaire/business-question.ejs
@@ -163,14 +163,15 @@
                     How would you describe your product/service concept?            
                 </p>
                 <div class="input-container" >
-                    <input type="radio" id="" name="service_concept" value=""/>
+                    <input type="radio" id="" name="service_concept" value="50"/>
                     <span for="">We’ve introduced a radical concept that will create a new market for itself.</span><br/>
-                    <input type="radio" id="" name="service_concept" value=""/>
+                    <input type="radio" id="" name="service_concept" value="75"/>
                     <span for="">We've introduced a disruptive concept that will create a stir in the existing market.</span><br/>
-                    <input type="radio" id="" name="service_concept" value=""/>
+                    <input type="radio" id="" name="service_concept" value="100"/>
                     <span for="">We’ve reintroduced a pre-existing concept with some new features.</span><br/>
-                    <input type="radio" id="" name="service_concept" value=""/>
-                    <span for="">We’ve reintroduced a pre-existing concept with negligible changes</span><br/>            </div>
+                    <input type="radio" id="" name="service_concept" value="125"/>
+                    <span for="">We’ve reintroduced a pre-existing concept with negligible changes</span><br/>            
+                </div>
             </div>
             <div class="question-box">
                 <p class="question-para">
